+++ docs/constants/navigation.js
@@ -67,7 +67,7 @@ const generalDirectories = fs
 
 // --- Navigation ---
 
-const home = [
+export const home = [
   makeSection('Get started', [
     makePage('get-started/introduction.mdx'),
     makePage('get-started/create-a-project.mdx'),
@@ -150,7 +150,7 @@ const home = [
   makeSection('More', [makePage('core-concepts.mdx'), makePage('faq.mdx')]),
 ];
 
-const general = [
+export const general = [
   makeSection('', [makePage('guides/overview.mdx')]),
   makeSection('Development process', [
     makePage('workflow/overview.mdx'),
@@ -371,7 +371,7 @@ const general = [
   ]),
 ];
 
-const eas = [
+export const eas = [
   makeSection(
     '',
     [
@@ -546,7 +546,7 @@ const eas = [
   ]),
 ];
 
-const learn = [
+export const learn = [
   makeSection('', [makePage('tutorial/overview.mdx')]),
   makeSection(
     'Expo tutorial',
@@ -666,7 +666,7 @@ const versionsReference = VERSIONS.reduce(
   {}
 );
 
-const reference = { ...versionsReference, latest: versionsReference['latest'] };
+export const reference = { ...versionsReference, latest: versionsReference['latest'] };
 
 export default {
   home,