+++ packages/expo-image-picker/android/src/main/java/expo/modules/imagepicker/contracts/ImageLibraryContract.kt
@@ -53,6 +53,8 @@ internal class ImageLibraryContract(
           }
         }
       )
+      .setOrderedSelection(input.options.orderedSelection)
+      .setDefaultTab(input.options.defaultTab.toDefaultTab())
       .build()
 
     if (input.options.allowsMultipleSelection) {