+++ packages/expo-ui/ios/SectionView.swift
@@ -17,8 +17,8 @@ struct SectionView: ExpoSwiftUI.View {
         Children().padding(EdgeInsets(top: 0, leading: 0, bottom: props.heightOffset, trailing: 0))
       }
     }
-
-    if #available(iOS 16.0, *) {
+    
+    if #available(iOS 16.0, tvOS 16.0, *) {
       form.scrollDisabled(true)
     } else {
       form